fixed compiler warnings

+++ src/model/demonlist/record.rs
@@ -251,7 +251,7 @@ impl FullRecord {
                     "footer": {
                         "text": format!("This record has been submitted by submitter #{}", self.submitter.map(|s|s.id).unwrap_or(1))
                     },
-                    "color": ((158 as f32 * progress) as i32) << 16 + ((224 as f32 * progress) as i32) << 8,
+                    "color": ((158_f32 * progress) as i32) << 16 + ((224_f32 * progress) as i32) << 8,
                     "author": {
                         "name": format!("{} (ID: {})", self.player.name, self.player.id),
                         "url": self.video
